refactor(services): add services into application folder
