+++ src/application/services/specialty.service.ts
