+++ src/application/services/treatment-type.service.spec.ts
