+++ src/application/services/treatment-type.service.ts
