--- src/modules/professionals/services/professionals.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { ProfessionalsService } from './professionals.service';
-
-describe('ProfessionalService', () => {
-	let service: ProfessionalsService;
-
-	beforeEach(async () => {
-		const module: TestingModule = await Test.createTestingModule({
-			providers: [ProfessionalsService],
-		}).compile();
-
-		service = module.get<ProfessionalsService>(ProfessionalsService);
-	});
-
-	it('should be defined', () => {
-		expect(service).toBeDefined();
-	});
-});
