--- src/modules/professionals/services/professionals.service.ts
@@ -1,4 +1,0 @@
-import { Injectable } from '@nestjs/common';
-
-@Injectable()
-export class ProfessionalsService {}
